[level 2] Title: 멀리 뛰기, Time: 0.18 ms, Memory: 76.4 MB -BaekjoonHub

+++ 프로그래머스/2/12914. 멀리 뛰기/멀리 뛰기.java
@@ -1,7 +1,7 @@
 class Solution {
     public long solution(int n) {
         if (n < 3) return n;
-        int[] jump = new int[n + 1];
+        long[] jump = new long[n + 1];
         jump[1] = 1;
         jump[2] = 2;
         for (int i = 3; i < n + 1; i++) {
